LPS-52822 Moving StagedTheme and StagedThemeImpl to model package along with StagedGroup and StagedGroupImpl
