+++ portal-impl/src/com/liferay/portal/staging/StagingImpl.java
@@ -87,6 +87,7 @@ import com.liferay.portal.model.LayoutRevision;
 import com.liferay.portal.model.Lock;
 import com.liferay.portal.model.Portlet;
 import com.liferay.portal.model.StagedModel;
+import com.liferay.portal.model.StagedTheme;
 import com.liferay.portal.model.User;
 import com.liferay.portal.model.WorkflowInstanceLink;
 import com.liferay.portal.security.auth.HttpPrincipal;
@@ -124,7 +125,6 @@ import com.liferay.portlet.documentlibrary.DuplicateFileException;
 import com.liferay.portlet.documentlibrary.FileExtensionException;
 import com.liferay.portlet.documentlibrary.FileNameException;
 import com.liferay.portlet.documentlibrary.FileSizeException;
-import com.liferay.portlet.layoutsadmin.lar.StagedTheme;
 
 import java.io.Serializable;
