+++ portal-impl/src/com/liferay/portlet/sitesadmin/lar/StagedGroupStagedModelDataHandler.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.lar.PortletDataContext;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.model.Group;
+import com.liferay.portal.model.StagedGroup;
 import com.liferay.portal.service.GroupLocalServiceUtil;
 
 import java.util.Map;
