--- portal-service/src/com/liferay/portlet/sitesadmin/lar/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
